pyproject.toml
[project]
name = "multimodal-injection-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.11.0",
    "bitsandbytes==0.45.4",
    "datasets>=2.14.4",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.7",
    "markupsafe==2.1.5",
    "matplotlib>=3.10.7",
    "openai>=2.6.1",
    "pillow>=11.3.0",
    "pyarrow==14.0.1",
    "python-dotenv>=1.2.1",
    "torch>=2.5.1",
    "torchaudio>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.66.5",
    "transformers>=4.57.1",
]

[tool.uv.extra-build-dependencies]
    flash-attn = ["torch"]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"

[[tool.uv.index]]
url = "https://pypi.org/simple"
